% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_qc.R
\name{align_sequences}
\alias{align_sequences}
\title{Align a set of DNA sequences}
\usage{
align_sequences(sequences, verbose = FALSE)
}
\arguments{
\item{sequences}{Character vector of DNA sequences}

\item{verbose}{Logical, print \code{DECIPHER::AlignSeqs()}'s progress output,
default \code{FALSE} (it's chatty by default -- fine for interactive use,
unwanted noise when called many times in a loop, e.g. from
\code{\link[=flag_barcode_gap_outliers]{flag_barcode_gap_outliers()}})}
}
\value{
A \code{Biostrings::DNAStringSet} of aligned sequences
}
\description{
Thin, namespaced wrapper over \code{Biostrings::DNAStringSet()} +
\code{DECIPHER::AlignSeqs()}.
}
