% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{classify_taxon_group}
\alias{classify_taxon_group}
\title{Classify BDQC taxonomic groups into finer display categories}
\usage{
classify_taxon_group(
  group_en,
  order = NA,
  class = NA,
  phylum = NA,
  kingdom = NA
)
}
\arguments{
\item{group_en}{Character vector, BDQC's coarse taxonomic group}

\item{order}{Character vector, taxonomic order (same length as \code{group_en})}

\item{class}{Character vector, taxonomic class (same length as \code{group_en})}

\item{phylum}{Character vector, taxonomic phylum (same length as \code{group_en})}

\item{kingdom}{Character vector, taxonomic kingdom (same length as \code{group_en})}
}
\value{
Character vector of finer taxonomic group labels
}
\description{
Reclassifies the coarse \code{group_en} column of the BDQC species list into
finer taxonomic categories used for figures (e.g. splitting mammals into
marine/terrestrial, arthropods into insects/crustaceans).
}
\examples{
classify_taxon_group("Mammals", order = "Cetacea")
classify_taxon_group("Arthropods", class = "Insecta")
}
