% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCGenomLandscape-package.R
\docType{package}
\name{QCGenomLandscape-package}
\alias{QCGenomLandscape}
\alias{QCGenomLandscape-package}
\title{QCGenomLandscape: Portrait of Genomic Data Availability for Quebec Species}
\description{
Tools to query BOLD Systems and NCBI for genomic sequence availability across the species of the Biodiversite Quebec (BDQC) species list, cross-reference conservation status (COSEPAC/LEMV), and build summary tables and figures.
}
\author{
\strong{Maintainer}: Steve Vissault \email{svissault01@gmail.com}

Authors:
\itemize{
  \item Steve Vissault \email{svissault01@gmail.com}
  \item Marie Pier Brochu
  \item Valérie Langlois
}

}
\keyword{internal}
