% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_qc.R
\name{align_sequences}
\alias{align_sequences}
\title{Align a set of DNA sequences}
\usage{
align_sequences(sequences)
}
\arguments{
\item{sequences}{Character vector of DNA sequences}
}
\value{
A \code{Biostrings::DNAStringSet} of aligned sequences
}
\description{
Thin, namespaced wrapper over \code{Biostrings::DNAStringSet()} +
\code{DECIPHER::AlignSeqs()}.
}
